Graph influence requirements
When a cell in the power table is clicked, show a graph of the
odds for the influence requirements associated with that row.
Each type of influence is graphed individually.

When mousing over the graph, draw a vertical track indicating the
number of draws, and odds for influence components at that draw
count along the left side of the graph.

Previously, if multiple lines in the decklist referenced the same card,
the number of instances of that card in the deck would be the number
on the last line.  Instead, the total should be the sum of the card
inclusion counts.  This is now the case.

A few Hooru cards from Set3 were incorrectly marked as Elysian cards.
This is now corrected.

Sort the rows of the draw probability table by the amount of power
required by the card with the influence requirement.