List power and influence sources, other minor improvements
Below the table of draw odds, list the names of all the cards
providing power and influence.  This will more clearly communicate
that favors, strangers and other cards which provide power and
influence when played are considered as sources.

There is now a back button in the upper left of the power and
influence graph.  Rather than clicking on the graph to return
to the table, now one clicks the back button.

Added an explanation of what the computed probabilities represent,
including the limitation that the requirements for playing sources
of power and influence aren't considered -- only the odds of having
those sources in hand are computed.

Rather than labelling the opening hand "7 draws", use labels
for the number of draws beyond the opening hand.  This will be
easier for people to understand and interpret.

Draw the graph influence labels after all the curves are drawn.
Previously, the influence icons could appear to be below other curves,
which looked weird.

Feln Stranger had incorrect influences.  Yeti Furflinger had the wrong
ultimate cost.  These are fixed.

Improve the HTML layout, assuming that the browser window is
at least 1280x720.