Graph all influences in deck
Rather than graphing the influences and power requirements related
to a specific row in the probability table, graph the influence
requirements of all the cards in the deck.  With this change, the
power requirements of cards are no longer part of the graph.

Graph curves start at a number of draws corresponding to the power
cost of the least expensive card in the deck with that influence
requirement.  This is useful to compare the probabilities of the
influence requirements with a number of draws at which they might become
relevant, which wasn't at all obvious previously -- as the entire
curve was previously drawn.

Added buttons above the graph and table for switching the view
from graph to table and back.  This replaces the action of
clicking on the table to go to the graph and the back button
on the graph.

Split epc.js into several components:

    epc-deck.js - data structures for cards and decks
    epc-graph.js - graph drawing
    epc-table.js - odds table generation
    epc-test.js - tests for odds computation