Regenerate graph when the influence icons finish loading
Previously, if the graph image was generated before the influence
icons have finished loading, the influence icons would be missing
from the graph, even after they had finished loading.  This was
particularly the case under Firefox, as it seems to be more prone
to loading images after javascript than other browsers.

Now the graph image is regenerated when the influence icons finish
loading.