Use Semantic-UI for buttons and modal dialogs
Added Semantic-UI to the project for better user interface components.
It allows us to add a modal dialog for importing a decklist, and then
use +/- buttons to modify the quantify of cards in the deck.  Individual
cards can now be added through a dialog with a dropdown used to select
from among all possible cards.  An "Export" button can now be used to
copy the current deck to the clipboard.

Browser local storage is used to make the deck persistent across
sessions.  This means that when you exit your browser and come back
to the power calculator, you'll have the same deck you were editing
the last time you were there.

When the odds are computed for situations where the number of cards
drawn is greater than the total number of cards in the deck, add
up all the influence sources in the deck to determine whether the
influence requirement can be satisfied.  (Previously, we were always
returning 100% odds, regardless of the influence sources actually
in the deck.  That was wrong.)

Previously, the order of the rows in the table could spontaneously
swap when changing the count of cards in the deck.  Now the ordering
is consistent, sorted first by power count, and then by card id
number.

Internet Explorer lacks Object.values, so we've now got a polyfill for
that.

Unfortuantely, with all these changes, the power calculator is no longer
function in Internet Explorer 9 or earlier.  However, all functionality
works correctly in IE10+, Chrome, Firefox, Safari and Edge.  (Luckily,
as of May 2018, IE versions earlier than IE10 account for only 1.3% of all
browsers in use, according to netmarketshare.com.)