Match the layout of Papa Capricorn's concept
Use a layout targeting the concept image created by Papa Capricorn.
Some additional graphStyle parameters have been added to the graph
to assist in matching that target.

Added CSS for Semantic-UI components to use fonts matching the
static elements of the design.

Used more agressive compression on the background image, reducing
page load time.  It was the largest single element contributing
to page load time.

Removed the default Semantic-UI font from the generated
CSS.  It was unnecessary, as we weren't using it, and could slow
the page loading.

Changed the fonts for the Power Odds Table to use a heavier font
for the header row and influence cost column and a lighter font
for the probabilities.

When the power and influence sources list is empty, refrain
from showing the "Power and influence sources:" label.

When the deck changes, the power and influence counts in the
influence panel are updated.

Revised the text in the About dialog for clarity.

Moved the About dialog to the top of the HTML file, which I believe
gives it more influence over Google indexing and AdSense associations.

Against my some of my reservations, I've switched to using "Turns"
in the power odds table, rather than using "Draws."  I don't like
that it is technically incorrect in the face of card draw effects,
but have come around for consistency with the influence graph and
ease of interpretation without doing arithmetic in your head.

Added "Lieutenant Relia" to card library.  Corrected "Forgotten Find"
to provide two wild influence, rather than just one wild influence.