Power odds table has 20 turn plus layout improvements
Add power odds table content up to 25 turns, and add a scrollbar
to see the additional content.

Rather than computing all of the odds for the table immediately
when the deck changes, compute the odds when the browser is idle.
This makes the UI much more responsive to changing card counts,
redrawing the graph immediately.

When hovering the mouse over a cell of the component odds table,
show a tooltip breaking down the odds for each component of the
influence associated with the cell.  (For example, a
Channel the Tempest deck will break down the cell into the
odds for eight power, and the odds for four primal influence
into individual probabilities.)

Changed the background color of the chart back to #000000.

Add labels along the edge of the graph documenting the axes.

Added a navigation menu in the upper right corner of the page
with links to other relevant pages on the site.

There is now an "About" link in the page header, rather than a
button in the deck editing panel.

Added a Clear button to reset the decklist to an empty deck.

The tabs selecting between the influence graph and power odds table
now have a help icon which can be hovered over to bring up a
description tooltip.

Separate the power cards in the decklist from the non-power cards.
In section for cards which aren't type power card type, highlight
the cards which are capable of providing influence or power.

Previously, the power odds table was sorted by the power and influence
of the primary casting cost of the card, and then each secondary
power and influence requirement had their rows following the primary
requirement.

Instead, now all costs are sorted by power and then influence,
regardless of whether they are primary or secondary requirements, and
regardless of which card they are attached to.

dist.sh can be used to place the components intended for
distribution into a dist/ subdirectory.  This includes some
minor preprocessing on the index.html to included a minified
version of the Javascript, rather than the readable and debuggable
version.