Show tooltips on influence graph
Draw dots along the curves in the influence graph at the position
where there is a card with power and influence corresponding
to that influence color and turn number.

When mousing over influence dots on the graph, show a tooltip
listing an exact percentage and cards relevant to that point
on the graph.

When hovering over the cost cell in the power odds table,
show a tooltip with the names of the cards which have that
cost.

In the power odds table, clarify the power requirement in the tooltip
by presenting the text as "3 Power - 82%" rather than as "3 - 82%".

Merged back changes to the about dialog from Papa Capricorn